compute highest of 3 integers

+++ 0x03-debugging/2-largest_number.c
@@ -8,9 +8,9 @@
  * Return: largest number
  */
 
-float largest_number(float a, float b, float c)
+int largest_number(int a, int b, int c)
 {
-	float largest;
+	int largest;
 
 	if (a >= b)
 	{
